challenges/web-example/requirements.txt
fastapi
uvicorn
pycryptodome
itsdangerous
